requirements.txt
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
distlib==0.3.6
dnspython==2.4.2
env==0.1.0
filelock==3.12.0
Flask==2.3.2
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.3
gradescope-utils==0.5.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
json2html==1.3.0
MarkupSafe==2.1.2
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.1
platformdirs==3.5.0
protobuf==4.21.12
psycopg2-binary==2.9.7
pymongo==4.5.0
PyMySQL==1.1.0
requests==2.30.0
SQLAlchemy==2.0.13
subprocess32==3.5.4
typing_extensions==4.5.0
urllib3==2.0.2
virtualenv==20.23.0
Werkzeug==2.3.4
